+++ MoreShipUpgrades/Misc/LategameConfiguration.cs
@@ -35,6 +35,7 @@ namespace MoreShipUpgrades.Misc
     public class LategameConfiguration : SyncedConfig2<LategameConfiguration>
     {
         #region Enabled
+        [field: SyncedEntryField] public SyncedEntry<bool> PARTICLE_INFUSER_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> SILVER_BULLETS_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> FUSION_MATTER_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> LONG_BARREL_ENABLED {  get; set; }
@@ -125,6 +126,7 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Initial Prices
+        [field: SyncedEntryField] public SyncedEntry<int> PARTICLE_INFUSER_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> SILVER_BULLETS_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> FUSION_MATTER_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> LONG_BARREL_PRICE { get; set; }
@@ -183,6 +185,10 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Attributes
+        [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_PRICES {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_OVERRIDE_NAME {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE { get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> PARTICLE_INFUSER_INCREMENTAL_TELEPORT_SPEED_INCREASE { get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> SILVER_BULLETS_OVERRIDE_NAME {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> FUSION_MATTER_PRICES {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> FUSION_MATTER_OVERRIDE_NAME {  get; set; }
@@ -508,6 +514,7 @@ namespace MoreShipUpgrades.Misc
         [field: SyncedEntryField] public SyncedEntry<string> ITEM_PROGRESSION_APPARATICE_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> ITEM_PROGRESSION_NO_PURCHASE_UPGRADES {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> ITEM_PROGRESSION_ALWAYS_SHOW_ITEMS {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> SILVER_BULLETS_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> FUSION_MATTER_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> LONG_BARREL_ITEM_PROGRESSION_ITEMS {  get; set; }
@@ -616,6 +623,7 @@ namespace MoreShipUpgrades.Misc
 
             topSection = LguConstants.OVERRIDE_NAMES_SECTION;
             OVERRIDE_UPGRADE_NAMES              = cfg.BindSyncedEntry(topSection, LguConstants.OVERRIDE_NAMES_ENABLED_KEY, LguConstants.OVERRIDE_NAMES_ENABLED_DEFAULT, LguConstants.OVERRIDE_NAMES_ENABLED_DESCRIPTION);
+            PARTICLE_INFUSER_OVERRIDE_NAME      = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_OVERRIDE_NAME_KEY, ParticleInfuser.UPGRADE_NAME);
             SILVER_BULLETS_OVERRIDE_NAME        = cfg.BindSyncedEntry(topSection, LguConstants.SILVER_BULLETS_OVERRIDE_NAME_KEY, SilverBullets.UPGRADE_NAME);
             FUSION_MATTER_OVERRIDE_NAME         = cfg.BindSyncedEntry(topSection, LguConstants.FUSION_MATTER_OVERRIDE_NAME_KEY, FusionMatter.UPGRADE_NAME);
             LONG_BARREL_OVERRIDE_NAME           = cfg.BindSyncedEntry(topSection, LguConstants.LONG_BARREL_OVERRIDE_NAME_KEY, LongBarrel.UPGRADE_NAME);
@@ -718,7 +726,19 @@ namespace MoreShipUpgrades.Misc
 
             #region Upgrades
 
-            #region Fast Encryption
+            #region Particle Infuser
+
+            topSection = ParticleInfuser.UPGRADE_NAME;
+            PARTICLE_INFUSER_ENABLED = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_ENABLED_KEY, LguConstants.PARTICLE_INFUSER_ENABLED_DEFAULT, LguConstants.PARTICLE_INFUSER_ENABLED_DESCRIPTION);
+            PARTICLE_INFUSER_PRICE = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_PRICE_KEY, LguConstants.PARTICLE_INFUSER_PRICE_DEFAULT);
+            PARTICLE_INFUSER_PRICES = cfg.BindSyncedEntry(topSection, BaseUpgrade.PRICES_SECTION, ParticleInfuser.DEFAULT_PRICES, BaseUpgrade.PRICES_DESCRIPTION);
+            PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE_KEY, LguConstants.PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE_DEFAULT, LguConstants.PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE_DESCRIPTION);
+            PARTICLE_INFUSER_INCREMENTAL_TELEPORT_SPEED_INCREASE = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_INCREMENTAL_TELEPORT_SPEED_INCREASE_KEY, LguConstants.PARTICLE_INFUSER_INCREMENTAL_TELEPORT_SPEED_INCREASE_DEFAULT, LguConstants.PARTICLE_INFUSER_INCREMENTAL_TELEPORT_SPEED_INCREASE_DESCRIPTION);
+            PARTICLE_INFUSER_ITEM_PROGRESSION_ITEMS = cfg.BindSyncedEntry(topSection, LguConstants.ITEM_PROGRESSION_ITEMS_KEY, LguConstants.ITEM_PROGRESSION_ITEMS_DEFAULT, LguConstants.ITEM_PROGRESSION_ITEMS_DESCRIPTION);
+
+            #endregion
+
+            #region Silver Bullets
 
             topSection = SilverBullets.UPGRADE_NAME;
             SILVER_BULLETS_ENABLED = cfg.BindSyncedEntry(topSection, LguConstants.SILVER_BULLETS_ENABLED_KEY, LguConstants.SILVER_BULLETS_ENABLED_DEFAULT, LguConstants.SILVER_BULLETS_ENABLED_DESCRIPTION);